Reorganize vector spaces and functions using new package and naming conventions.
