+++ src/main/scala/org/apache/spark/mllib/optimization/tfocs/ProxCapableFunction.scala
@@ -17,8 +17,6 @@
 
 package org.apache.spark.mllib.optimization.tfocs
 
-import org.apache.spark.mllib.linalg.{ DenseVector, Vector, Vectors }
-
 /**
  * A trait for prox capable functions which support efficient proximity minimization, as expressed
  * by the proximity operator:
@@ -49,118 +47,4 @@ trait ProxCapableFunction[X] {
 
   /** Evaluate the function h(x) at x. Does not perform proximity minimization. */
   def apply(x: X): Double
-}
-
-/**
- * The proximity operator for constant zero.
- *
- * NOTE In matlab tfocs this functionality is implemented in prox_0.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/prox_0.m]]
- */
-class ProxZero extends ProxCapableFunction[DenseVector] {
-
-  override def apply(z: DenseVector, t: Double, mode: ProxMode): ProxValue[DenseVector] =
-    ProxValue(Some(0.0), Some(z))
-
-  override def apply(x: DenseVector): Double = 0.0
-}
-
-/**
- * The proximity operator for the L1 norm, with a scale 'q' applied.
- *   q * ||x||_1
- *
- * @param q A positive scalar factor to multiply by the computed L1 norm. Must be > 0.
- *
- * NOTE In matlab tfocs this functionality is implemented in prox_l1.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/prox_l1.m]]
- */
-class ProxL1(q: Double) extends ProxCapableFunction[DenseVector] {
-
-  require(q > 0)
-
-  override def apply(z: DenseVector, t: Double, mode: ProxMode): ProxValue[DenseVector] = {
-    // NOTE DenseVectors are assumed here (not sparse safe).
-    val shrinkage = q * t
-    val minimizer = shrinkage match {
-      case 0.0 => z
-      case _ => new DenseVector(z.values.map(z_i =>
-        z_i * (1.0 - math.min(shrinkage / math.abs(z_i), 1.0))))
-    }
-    val f = if (mode.f) Some(apply(minimizer)) else None
-    ProxValue(f, Some(minimizer))
-  }
-
-  override def apply(x: DenseVector): Double = q * Vectors.norm(x, 1)
-}
-
-/**
- * A projection onto the nonnegative orthant, implemented using a zero/infinity indicator function.
- *
- * NOTE In matlab tfocs this functionality is implemented in proj_Rplus.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/proj_Rplus.m]]
- */
-class ProjRPlus extends ProxCapableFunction[DenseVector] {
-
-  override def apply(z: DenseVector, t: Double, mode: ProxMode): ProxValue[DenseVector] = {
-
-    val minimizer = if (mode.minimizer) {
-      // NOTE DenseVectors are assumed here (not sparse safe).
-      Some(new DenseVector(z.values.map(math.max(_, 0.0))))
-    } else {
-      None
-    }
-
-    ProxValue(Some(0.0), minimizer)
-  }
-
-  override def apply(x: DenseVector): Double =
-    if (x.values.min < 0.0) Double.PositiveInfinity else 0.0
-}
-
-/**
- * A projection onto a simple box defined by upper and lower limits on each vector element,
- * implemented using a zero/infinity indicator function.
- *
- * @param l A vector describing the box's lower bound. If x is within the box then x_i >= l_i for
- *        all i.
- * @param u A vector describing the box's upper bound. If x is within the box then x_i <= u_i for
- *        all i.
- *
- * NOTE In matlab tfocs this functionality is implemented in proj_box.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/proj_box.m]]
- */
-class ProjBox(l: DenseVector, u: DenseVector) extends ProxCapableFunction[DenseVector] {
-
-  override def apply(z: DenseVector, t: Double, mode: ProxMode): ProxValue[DenseVector] = {
-
-    val minimizer = if (mode.minimizer) {
-      // NOTE DenseVectors are assumed here (not sparse safe).
-      val ret = new Array[Double](z.size)
-      var i = 0
-      while (i < ret.size) {
-        // Bound each element using the lower and upper limit for that element.
-        ret(i) = math.min(u(i), math.max(l(i), z(i)))
-        i += 1
-      }
-      Some(new DenseVector(ret))
-    } else {
-      None
-    }
-
-    ProxValue(Some(0.0), minimizer)
-  }
-
-  override def apply(x: DenseVector): Double = {
-    // NOTE DenseVectors are assumed here (not sparse safe).
-    var ret = 0.0
-    var i = 0
-    while (i < x.size) {
-      // If an element is outside of that element's bounds, return infinity.
-      if (x(i) > u(i) || x(i) < l(i)) {
-        ret = Double.PositiveInfinity
-      }
-      i += 1
-    }
-    ret
-  }
 }
