+++ src/main/scala/org/apache/spark/mllib/optimization/tfocs/SmoothFunction.scala
@@ -17,11 +17,6 @@
 
 package org.apache.spark.mllib.optimization.tfocs
 
-import org.apache.spark.mllib.linalg.{ Vector, Vectors }
-import org.apache.spark.mllib.optimization.tfocs.DVectorFunctions._
-import org.apache.spark.mllib.optimization.tfocs.VectorSpace._
-import org.apache.spark.storage.StorageLevel
-
 /**
  * A trait for smooth functions, with support for evaluating the function and computing its
  * gradient.
@@ -46,134 +41,4 @@ trait SmoothFunction[X] {
    * Evaluates the function on vector x.
    */
   def apply(x: X): Double = apply(x, Mode(f = true, g = false)).f.get
-}
-
-/**
- * The squared error function applied to a DVector, with a constant factor of 0.5.
- *
- * @param x0 The base vector against which the squared error difference is computed.
- *
- * NOTE In matlab tfocs this functionality is implemented in smooth_quad.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/smooth_quad.m]]
- */
-class SmoothQuad(x0: DVector) extends SmoothFunction[DVector] {
-
-  if (x0.getStorageLevel == StorageLevel.NONE) {
-    x0.cache()
-  }
-
-  override def apply(x: DVector, mode: Mode): Value[DVector] = {
-
-    // Compute the squared error gradient (just the difference between vectors).
-    val g = x.diff(x0)
-
-    // If both f and g are requested then g will be read twice, so cache it.
-    if (mode.f && mode.g) g.cache()
-
-    val f = if (mode.f) {
-      // Compute the squared error.
-      // TODO If f is required but not g, then performance might be improved by reimplementing as
-      // a single aggregate using 'x' and 'x0' without an intermediate 'g' DVector, which breaks
-      // per-element pipelining.
-      Some(g.aggregate(0.0)((sum, y) => sum + math.pow(Vectors.norm(y, 2), 2), _ + _) / 2.0)
-    } else {
-      None
-    }
-    Value(f, Some(g))
-  }
-}
-
-/**
- * The Huber loss function applied to a DVector.
- *
- * @param x0 The vector against which loss should be computed.
- * @param tau The huber loss parameter.
- *
- * NOTE In matlab tfocs this functionality is implemented in smooth_huber.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/smooth_huber.m]]
- */
-class SmoothHuber(x0: DVector, tau: Double) extends SmoothFunction[DVector] with Serializable {
-
-  if (x0.getStorageLevel == StorageLevel.NONE) {
-    x0.cache()
-  }
-
-  override def apply(x: DVector, mode: Mode): Value[DVector] = {
-
-    val diff = x.diff(x0)
-
-    // If both f and g are requested then diff will be read twice, so cache it.
-    if (mode.f && mode.g) diff.cache()
-
-    val f = if (mode.f) {
-      // TODO If f is required but not g, then performance might be improved by reimplementing as
-      // a single aggregate using 'x' and 'x0' without an intermediate 'diff' DVector, which breaks
-      // per-element pipelining.
-      Some(diff.aggregateElements(0.0)(
-        seqOp = (sum, diff_i) => {
-          // Find the huber loss, corresponding to the adjusted l2 loss when the magnitude is <= tau
-          // and to the adjusted l1 loss when the magnitude is > tau.
-          val huberValue = if (math.abs(diff_i) <= tau) {
-            0.5 * diff_i * diff_i / tau
-          } else {
-            math.abs(diff_i) - tau / 2.0
-          }
-          sum + huberValue
-        },
-        combOp = _ + _))
-    } else {
-      None
-    }
-
-    val g = if (mode.g) {
-      // Compute the huber loss gradient elementwise.
-      Some(diff.mapElements(diff_i => diff_i / math.max(math.abs(diff_i), tau)))
-    } else {
-      None
-    }
-
-    Value(f, g)
-  }
-}
-
-/**
- * The log likelihood logistic loss function applied to a DVector.
- *
- * @param y The observed values, labeled as binary 0/1.
- * @param mu The variable values.
- *
- * NOTE In matlab tfocs this functionality is implemented in smooth_logLLogistic.m.
- * @see [[https://github.com/cvxr/TFOCS/blob/master/smooth_logLLogistic.m]]
- */
-class SmoothLogLLogistic(y: DVector) extends SmoothFunction[DVector] with Serializable {
-
-  if (y.getStorageLevel == StorageLevel.NONE) {
-    y.cache()
-  }
-
-  override def apply(mu: DVector, mode: Mode): Value[DVector] = {
-
-    val f = if (mode.f) {
-      // TODO Performance might be improved by reimplementing as a single aggregate rather than
-      // mapping through an intermediate DVector and summing, which breaks per-element pipelining.
-      Some(y.zipElements(mu, (y_i, mu_i) => {
-        val yFactor = if (mu_i > 0.0) y_i - 1.0 else if (mu_i < 0.0) y_i else 0.0
-        yFactor * mu_i - math.log1p(math.exp(-math.abs(mu_i)))
-      }).sum)
-    } else {
-      None
-    }
-
-    val g = if (mode.g) {
-      Some(y.zipElements(mu, (y_i, mu_i) => {
-        // Compute the log logistic loss gradient elementwise.
-        val muFactor = if (mu_i > 0.0) 1.0 else math.exp(mu_i)
-        y_i - muFactor / (1.0 + math.exp(-math.abs(mu_i)))
-      }))
-    } else {
-      None
-    }
-
-    Value(f, g)
-  }
 }
