+++ src/test/scala/org/apache/spark/mllib/optimization/tfocs/LinearOperatorSuite.scala
@@ -22,6 +22,8 @@ import org.scalatest.FunSuite
 import org.apache.spark.SparkException
 import org.apache.spark.mllib.linalg.{ DenseVector, Vectors }
 import org.apache.spark.mllib.optimization.tfocs.DVectorFunctions._
+import org.apache.spark.mllib.optimization.tfocs.fs.vector.dvector.LinopMatrix
+import org.apache.spark.mllib.optimization.tfocs.fs.dvector.vector.{ LinopMatrix => LinopMatrixAdjoint }
 import org.apache.spark.mllib.util.MLlibTestSparkContext
 
 class LinearOperatorSuite extends FunSuite with MLlibTestSparkContext {
@@ -29,9 +31,9 @@ class LinearOperatorSuite extends FunSuite with MLlibTestSparkContext {
   lazy val matrix = sc.parallelize(Array(Vectors.dense(1.0, 2.0, 3.0),
     Vectors.dense(4.0, 5.0, 6.0)), 2)
 
-  test("DenseVectorToDVectorLinOp multiplies properly") {
+  test("LinopMatrix multiplies properly") {
 
-    val f = new DenseVectorToDVectorLinOp(matrix)
+    val f = new LinopMatrix(matrix)
     val x = Vectors.dense(7.0, 8.0, 9.0).toDense
     val result = f(x)
     val expectedResult = Vectors.dense(1 * 7 + 2 * 8 + 3 * 9, 4 * 7 + 5 * 8 + 6 * 9)
@@ -39,18 +41,18 @@ class LinearOperatorSuite extends FunSuite with MLlibTestSparkContext {
       "should return the correct product")
   }
 
-  test("TransposedDenseVectorToDVectorLinOp multiplies properly") {
+  test("LinopMatrixAdjoint multiplies properly") {
 
-    val f = new TransposedDenseVectorToDVectorLinOp(matrix)
+    val f = new LinopMatrixAdjoint(matrix)
     val y = sc.parallelize(Array(Vectors.dense(5.0).toDense, Vectors.dense(6.0).toDense), 2)
     val result = f(y)
     val expectedResult = Vectors.dense(1 * 5 + 4 * 6, 2 * 5 + 5 * 6, 3 * 5 + 6 * 6)
     assert(result == expectedResult, "should return the correct product")
   }
 
-  test("TransposedDenseVectorToDVectorLinOp checks for mismatched partition vectors") {
+  test("LinopMatrixAdjoint checks for mismatched partition vectors") {
 
-    val f = new TransposedDenseVectorToDVectorLinOp(matrix)
+    val f = new LinopMatrixAdjoint(matrix)
     val y = sc.parallelize(Array(Vectors.dense(5.0, 6.0).toDense, Vectors.zeros(0).toDense), 2)
     intercept[SparkException] {
       f(y)
